+++ src/components/Home/Suggestions.vue
@@ -1,17 +1,14 @@
 <script setup>
 import { VueperSlide, VueperSlides } from 'vueperslides'
 import 'vueperslides/dist/vueperslides.css'
+import { useHomeStore } from '@/stores/home.js'
 
-defineProps({
-    suggestions: {
-        required: true,
-    },
-})
+const homeStore = useHomeStore()
 </script>
 
 <template>
-    <vueper-slides :bullets="false" :gap="2" :infinite="true" :slide-ratio="1 / 4" :visible-slides="3" autoplay class="no-shadow h-full" duration="2000" fixed-height="300px">
-        <vueper-slide v-for="book in suggestions" :key="book.id" :image="book.images" class="rounded-2xl" />
+    <vueper-slides :infinite="true" duration="2000" autoplay class="no-shadow h-full" :visible-slides="3" :gap="2" :bullets="false" :slide-ratio="1 / 4" fixed-height="300px">
+        <vueper-slide v-for="book in homeStore.suggestions" :key="book.id" :image="book.images" class="rounded-2xl" />
     </vueper-slides>
 </template>
