+++ src/components/Profile/UserInfo.vue
@@ -1,30 +1,46 @@
 <script setup>
-import { computed, reactive, ref } from 'vue'
+import { ref } from 'vue'
 import { useProfileStore } from '@/stores/profile.js'
+import { formatDate } from 'compatx'
 import editProfileApi from '@/api/editProfile.js'
-import { formatDate, removeNullKeys } from '@/utilities.js'
 import ErrorPopup from '@/components/common/ErrorPopup.vue'
 
 const profileStore = useProfileStore()
-const user = computed(() => profileStore.user.user)
-const profile_pic = computed(() => (user.value.profile_picture !== null ? user.value.profile_picture : 'src/assets/images/cat.jpg'))
-const isFlipped = ref(false)
-const errorMessage = ref('')
+const user = profileStore.user.user
+const profile_pic = user.profile_picture !== null ? user.profile_picture : 'src/assets/images/cat.jpg'
 
-const data = reactive({
+const isFlipped = ref(false)
+
+const flipCard = () => {
+    isFlipped.value = !isFlipped.value
+}
+const data = ref({
     name: null,
     username: null,
     date_of_birth: null,
     gender: null,
     profile_picture: null,
 })
+console.log(user)
 function handleFileChange(event) {
-    data.profile_picture = event.target.files[0]
+    data.value.profile_picture = event.target.files[0]
 }
+
+function removeNullKeys(obj) {
+    for (const key in obj) {
+        if (obj[key] === null) {
+            delete obj[key]
+        }
+    }
+}
+
+const errorMessage = ref('')
+
 async function updateProfile() {
     try {
-        removeNullKeys(data)
+        removeNullKeys(data.value)
         const response = await editProfileApi.edit(data.value)
+        console.log(response.data)
         if (response.status === 200) {
             profileStore.user = response.data.user
             isFlipped.value = false
@@ -39,9 +55,6 @@ async function updateProfile() {
     } finally {
         //
     }
-}
-const flipCard = () => {
-    isFlipped.value = !isFlipped.value
 }
 </script>
 
@@ -83,7 +96,7 @@ const flipCard = () => {
                         <h1 class="text-2xl font-extra-light">Details</h1>
 
                         <p class="text-dark_blue inline">Joined in:</p>
-                        <span class="ml-4">{{ formatDate(user.created_at) }}</span>
+                        <span class="mr-4">{{ formatDate(user.created_at) }}</span>
                         <br />
                         <p class="mt-2 text-gray-600">
                             Favorite <span class="text-dark_blue text-lg">Genres</span>: <br />
